Upversion scala 213 (#1317)
* wip upgrading to scala 2.13.5; fix scanamo dependency

* wip; upgrade content-api-client-aws now 2.13 version is available and a bunch of other dependencies

* Lots of upversioning, add dependencyTree

* Correct imports, move to "with Logging"

* More replace logging, playing with dynamo db client for scanamo

* More debugging

* Lint for upgrade

* Build credentials for new awssdk package similarly to old

* Use new create methods for creds providers

* Change to new Scanamo operation and key format

* Change to non-deprecated methods

* Add formatters for Trail and its child types

* Fix type

* Lint

* Revert accidental change?

* Add brackets for function call

* Convert deprecated function

* Get StringEscapeUtils from different apache library (refactoring change)

* Remove unused import from Twirl imports

* Use Logging trait consistently

* Revert whitespace change

* Resolve deprecation warnings: use new asJavaCollection

* Compare strings better - do not assume parameter order

* Re-order json to make tests happy

* Use logging.Logger everywhere

* Upversion play to 2.8.0 to match org.scalikejdbc

* Lint and Snyk-lint

* Lint and Snyk-lint

+++ app/metrics/FrontendMetrics.scala
@@ -29,7 +29,7 @@ case class FrontendStatisticSet(metric: FrontendMetric, datapoints: List[DataPoi
   lazy val sampleCount: Double = datapoints.size
   lazy val maximum: Double = Try(datapoints.maxBy(_.value).value).getOrElse(0L).toDouble
   lazy val minimum: Double = Try(datapoints.minBy(_.value).value).getOrElse(0L).toDouble
-  lazy val sum: Double = datapoints.map(_.value).sum
+  lazy val sum: Double = datapoints.map(_.value).sum.toDouble
   lazy val average: Double =
     Try(sum / sampleCount).toOption.getOrElse(0L)
